kiosk_app/requirements.txt
python-socketio>=5.8.0
psutil>=5.9.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
loguru>=0.5.3 